Move Tailwind config injection to HydeFront facade
This is more coupled to HydeFront than general Asset services
