+++ packages/framework/resources/views/layouts/styles.blade.php
@@ -11,7 +11,7 @@
 {{-- Dynamic TailwindCSS Play CDN --}}
 @if(config('hyde.use_play_cdn', false))
     <script src="https://cdn.tailwindcss.com?plugins=typography"></script>
-    <script>tailwind.config = { {!! Asset::injectTailwindConfig() !!} }</script>
+    <script>tailwind.config = { {!! HydeFront::injectTailwindConfig() !!} }</script>
     <script>console.warn('The HydePHP TailwindCSS Play CDN is enabled. This is for development purposes only and should not be used in production.', 'See https://hydephp.com/docs/1.x/managing-assets');</script>
 @endif
