+++ packages/framework/tests/Unit/Facades/HydeFrontFacadeTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Unit\Facades;
 
+use Hyde\Hyde;
 use Hyde\Facades\HydeFront;
 use Hyde\Testing\UnitTestCase;
 
@@ -12,6 +13,8 @@ use Hyde\Testing\UnitTestCase;
  */
 class HydeFrontFacadeTest extends UnitTestCase
 {
+    protected static bool $needsKernel = true;
+
     // Todo: Check the version is correct? (When running in monorepo)
 
     public function testVersionReturnsString()
@@ -35,5 +38,25 @@ class HydeFrontFacadeTest extends UnitTestCase
     {
         $expected = 'https://cdn.jsdelivr.net/npm/hydefront@v3.4/dist/invalid';
         $this->assertSame($expected, HydeFront::cdnLink('invalid'));
+    }
+
+    public function testInjectTailwindConfigReturnsExtractedTailwindConfig()
+    {
+        $config = HydeFront::injectTailwindConfig();
+
+        $this->assertIsString($config);
+        $this->assertStringContainsString("darkMode: 'class'", $config);
+        $this->assertStringContainsString('theme: {', $config);
+        $this->assertStringContainsString('extend: {', $config);
+        $this->assertStringContainsString('typography: {', $config);
+        $this->assertStringNotContainsString('plugins', $config);
+    }
+
+    public function testInjectTailwindConfigHandlesMissingConfigFileGracefully()
+    {
+        rename(Hyde::path('tailwind.config.js'), Hyde::path('tailwind.config.js.bak'));
+        $this->assertIsString(HydeFront::injectTailwindConfig());
+        $this->assertSame('', HydeFront::injectTailwindConfig());
+        rename(Hyde::path('tailwind.config.js.bak'), Hyde::path('tailwind.config.js'));
     }
 }
